+++ src/twg2/parser/textParser/TextParser.java
@@ -4,6 +4,8 @@ import java.io.BufferedReader;
 import java.io.Closeable;
 import java.io.PushbackReader;
 
+import twg2.parser.textStream.LineCounter;
+
 /** A buffered reader like class that reads lines and also allows the current line
  * to be fully or partially unread.
  * @see PushbackReader
@@ -12,6 +14,12 @@ import java.io.PushbackReader;
  * @since 2013-12-10
  */
 public interface TextParser extends LineReader, TextParserConditionals, Closeable {
+
+	/**
+	 * @return get line number counter which tracks the current line and column number position of this text parser
+	 */
+	public LineCounter getLineNumbers();
+
 
 	/**
 	 * @return the remainder of the buffer's current line of text, excluding the ending '\n'