+++ src/twg2/parser/textStream/CharsLineSupplier.java
@@ -13,8 +13,8 @@ public class CharsLineSupplier extends LineSupplier implements Supplier<char[]> 
 	/**
 	 * @see LineSupplier#LineSupplier(String, int, int, boolean, boolean, boolean, boolean)
 	 */
-	public CharsLineSupplier(String str, int off, int len, boolean treatEmptyLineAsLine, boolean treatEolNewlineAsTwoLines, boolean includeNewlinesAtEndOfReturnedLines, boolean collapseNewlinesIntoOneChar) {
-		super(str, off, len, treatEmptyLineAsLine, treatEolNewlineAsTwoLines, includeNewlinesAtEndOfReturnedLines, collapseNewlinesIntoOneChar);
+	public CharsLineSupplier(String str, int off, int len, boolean includeEmptyLines, boolean treatEolNewlineAsTwoLines, boolean includeNewlinesAtEndOfReturnedLines, boolean collapseNewlinesIntoOneChar) {
+		super(str, off, len, includeEmptyLines, treatEolNewlineAsTwoLines, includeNewlinesAtEndOfReturnedLines, collapseNewlinesIntoOneChar);
 	}