+++ src/twg2/parser/textStream/LineCounter.java
@@ -12,29 +12,31 @@ import twg2.collections.primitiveCollections.IntListSorted;
  * @since 2016-09-10
  */
 public class LineCounter {
-	private IntListSorted lineStartOffsets;
-	private int off;
+	private IntListSorted lineOffsets;
+	private int curOffset;
+	private int maxReadPos;
 	private int curLineNum;
-	private char prevChar;
 
 
 	/** Create a line counter starting at the first character of the first line
 	 * @param startLineOffset absolute start character offset, this is the first line offset
 	 */
 	public LineCounter(int startLineOffset) {
-		this.lineStartOffsets = new IntListSorted();
-		this.off = startLineOffset;
+		this.lineOffsets = new IntListSorted();
+		this.curOffset = startLineOffset;
+		this.maxReadPos = startLineOffset;
 		this.curLineNum = 0;
-		this.lineStartOffsets.add(startLineOffset);
+		this.lineOffsets.add(startLineOffset);
 	}
 
 
 	/** Create a 0-based line counter given a list of line numbers
 	 */
 	public LineCounter(Collection<? extends Integer> prevLineStartOffsets) {
-		this.lineStartOffsets = IntListSorted.of(prevLineStartOffsets);
-		this.off = this.lineStartOffsets.getLast();
-		this.curLineNum = this.lineStartOffsets.size();
+		this.lineOffsets = IntListSorted.of(prevLineStartOffsets);
+		this.curOffset = this.lineOffsets.size() > 0 ? this.lineOffsets.getLast() : 0;
+		this.maxReadPos = this.curOffset;
+		this.curLineNum = this.lineOffsets.size() > 0 ? this.lineOffsets.size() - 1 : 0;
 	}
 
 
@@ -42,13 +44,36 @@ public class LineCounter {
 	 * @return the current line number (0-based)
 	 */
 	public int read(char ch) {
-		if(prevChar == '\n') {
-			lineStartOffsets.add(off);
-			curLineNum++;
+		// reading new characters
+		if(curOffset >= maxReadPos) {
+			curOffset++;
+			maxReadPos++;
+			if(ch == '\n') {
+				lineOffsets.add(curOffset);
+				curLineNum++;
+			}
 		}
-		prevChar = ch;
-		off++;
+		// re-reading characters after unread()
+		else {
+			curOffset++;
+			int nextLineOffset = lineOffsets.size() > curLineNum + 1 ? lineOffsets.get(curLineNum + 1) : -1;
+			if(nextLineOffset > -1 && curOffset >= nextLineOffset) {
+				curLineNum++;
+			}
+		}
 		return curLineNum;
+	}
+
+
+	public void unread(int count) {
+		if(this.curOffset - count < 0) { throw new IndexOutOfBoundsException("cannot unread " + count + " from " + this.curOffset + ", offset must remain greater than or equal to 0"); }
+		this.curOffset -= count;
+		this.curLineNum = getLineNumber(this.curOffset);
+	}
+
+
+	public int getLineNumber() {
+		return this.curLineNum;
 	}
 
 
@@ -57,7 +82,7 @@ public class LineCounter {
 	 * @return the 0-based line number of the line containing the character at {@code offset}
 	 */
 	public int getLineNumber(int offset) {
-		int idx = this.lineStartOffsets.binarySearch(offset);
+		int idx = this.lineOffsets.binarySearch(offset);
 		idx = idx < 0 ? -idx - 2 : idx;
 		return idx;
 	}
@@ -68,33 +93,30 @@ public class LineCounter {
 	 * @return the absolute character offset (0-based) of the first character of line number {@code lineIndex}
 	 */
 	public int getLineOffset(int lineIndex) {
-		return this.lineStartOffsets.get(lineIndex);
+		return this.lineOffsets.get(lineIndex);
 	}
 
 
 	/**
 	 * @return the number of lines this line counter has encountered
 	 */
-	public int size() {
-		return this.lineStartOffsets.size();
+	public int lineCount() {
+		return this.lineOffsets.size();
 	}
 
 
 	/** Indicates that input is complete and if the last char was a line ending, count it, and return the line numbers
 	 * @return this line counter's internal list of line number offsets based on the chars passed to {@link #read(char)} and the initial constructor offset.
-	 * NOTE: should not modify this list, if you'd like a copy, call {@link IntListSorted#copy()} on the return value
+	 * NOTE: you should not modify the returned list, if you'd like a copy, call {@link IntListSorted#copy()} on the return value
 	 */
 	public IntListSorted getRawCompletedLineOffsets() {
-		if(this.prevChar == '\n') {
-			this.lineStartOffsets.add(this.off);
-		}
-		return this.lineStartOffsets;
+		return this.lineOffsets;
 	}
 
 
 	@Override
 	public String toString() {
-		return "LineCounter: { currentLineNum: " + this.curLineNum + ", lines: " + this.lineStartOffsets + " }";
+		return "LineCounter: { currentLineNum: " + this.curLineNum + ", lines: " + this.lineOffsets + " }";
 	}
 
 }