+++ test/twg2/parser/textStream/test/LineCounterTest.java
@@ -17,7 +17,7 @@ public class LineCounterTest {
 	@Test
 	public void givenLines() {
 		LineCounter lc = new LineCounter(Arrays.asList(0, 12, 22, 62, 87, 88));
-		Assert.assertEquals(6, lc.size());
+		Assert.assertEquals(6, lc.lineCount());
 		assertAll(0, lc.getLineNumber(0), lc.getLineNumber(1), lc.getLineNumber(9), lc.getLineNumber(10), lc.getLineNumber(11));
 		assertAll(1, lc.getLineNumber(12), lc.getLineNumber(13), lc.getLineNumber(20), lc.getLineNumber(21));
 		assertAll(4, lc.getLineNumber(87));
@@ -26,27 +26,51 @@ public class LineCounterTest {
 
 
 	@Test
+	public void rewind() {
+		char[] chs = (
+			"start" + "\n" + // 6 chars
+			"A. len 11" + "\r\n" + // 11 chars
+			"B. len 9" + "\n" // 9 chars
+		).toCharArray();
+
+		LineCounter lc = new LineCounter(0);
+		for(int i = 0; i < 7; i++) { lc.read(chs[i]); } // read through first char of second line 'A'
+		Assert.assertEquals(1, lc.getLineNumber());
+		lc.unread(1);
+		Assert.assertEquals(1, lc.getLineNumber());
+		lc.unread(1);
+		Assert.assertEquals(0, lc.getLineNumber());
+		lc.unread(1);
+		Assert.assertEquals(0, lc.getLineNumber());
+		lc.read('_'); // the char doesn't matter when re-reading after unread()
+		Assert.assertEquals(0, lc.getLineNumber());
+		lc.read('_');
+		Assert.assertEquals(1, lc.getLineNumber());
+	}
+
+
+	@Test
 	public void addText() {
 		String lns = "start" + "\n" +
-				"1. len 11" + "\r\n" +
-				"2. len 9" + "\n";
-		int off = 6;
+				"A. len 11" + "\r\n" +
+				"B. len 9" + "\n";
 		// correct offsets
-		IntListSorted lnStarts = IntListSorted.of(0, off, 17, 26);
+		IntListSorted lnStarts = IntListSorted.of(0, 5, 16, 25);
 
-		LineCounter lc = new LineCounter(Arrays.asList(0, off));
+		LineCounter lc = new LineCounter(Arrays.asList(0, 5));
+		// start offset at the first character of the second line
 		char[] chs = lns.toCharArray();
-		int prevLine = -1;
+		int prevLine = 1;
 		// pass each char to the line counter and check that the line number of that character matches the expected ones from the known correct list
-		for(int i = off; i < chs.length; i++) {
-			int line = lc.read(chs[i]);
+		for(int i = 6; i < chs.length; i++) {
+			char ch = chs[i];
+			int line = lc.read(ch);
 
 			if(line != prevLine) {
-				int lnNum = lnStarts.binarySearch(i) + 1;
-				Assert.assertEquals("char " + i + " '" + chs[i] + "'", lnNum, line);
+				int lnNum = lnStarts.binarySearch(i);
+				Assert.assertEquals("char " + i + " '" + ch + "'", lnNum, line);
+				prevLine = line;
 			}
-
-			prevLine = line;
 		}
 
 		// check the completed list of offsets